remove more legacy code
